Rename MiniCPM-V package to minicpmv4_6
